feat: reorganize folders structure and extract background with larvae for future use
