+++ lib/src/common/widgets/background.dart
@@ -1,0 +1,62 @@
+import 'dart:ui';
+
+import 'package:flutter/material.dart';
+import 'package:larvixon_frontend/src/common/widgets/petri_dish/petri_dish.dart';
+
+class Background extends StatelessWidget {
+  const Background({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        gradient: LinearGradient(
+          colors: [
+            Theme.of(context).colorScheme.primary,
+            Theme.of(context).colorScheme.secondary,
+          ],
+          begin: Alignment.topLeft,
+          end: Alignment.bottomRight,
+        ),
+      ),
+    );
+  }
+}
+
+class BackgroundWithLarvae extends StatelessWidget {
+  final bool useBlur;
+  final int larvaeCount;
+  const BackgroundWithLarvae({
+    super.key,
+    this.useBlur = true,
+    this.larvaeCount = 3,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      fit: StackFit.expand,
+      children: [
+        const Background(),
+        PetriDish(larvaeCount: larvaeCount),
+        if (useBlur)
+          BackdropFilter(
+            filter: ImageFilter.blur(sigmaX: 5, sigmaY: 5),
+            child: Container(
+              decoration: BoxDecoration(
+                gradient: LinearGradient(
+                  colors: [
+                    Colors.black.withValues(alpha: 0.4),
+                    Colors.transparent,
+                  ],
+                  begin: Alignment.topCenter,
+                  end: Alignment.bottomCenter,
+                ),
+              ),
+              child: Container(color: Colors.transparent),
+            ),
+          ),
+      ],
+    );
+  }
+}
