--- lib/src/landing/presentation/petri_dish_painter.dart
@@ -1,1 +1,0 @@
-
